feat: #2064 integrate @reapit/connect-session (rename from connect-utils): update private-route-wrapper
