+++ packages/connect-session/src/react/index.ts
@@ -1,6 +1,7 @@
 import React from 'react'
 import { ReapitConnectHook, ReapitConnectSession } from '../types'
 import { ReapitConnectBrowserSession } from '../browser'
+;(window as any).a = React.useState
 
 export const useReapitConnect = (reapitConnectBrowserSession: ReapitConnectBrowserSession): ReapitConnectHook => {
   const [connectSession, setConnectSession] = React.useState<ReapitConnectSession | null>(null)
