+++ packages/cra-external-template/template/src/components/pages/login.tsx
@@ -1,47 +1,58 @@
 import * as React from 'react'
-import { Redirect } from 'react-router-dom'
+import { render } from '@testing-library/react'
+import { shallow } from 'enzyme'
+import { Button } from '@reapit/elements'
+import { ReapitConnectBrowserSessionInstance } from '../../../core/connect-session'
+import { createBrowserHistory } from 'history'
+import { AuthContext } from '../../../context'
+import { mockContext } from '../../../context/__mocks__/mock-context'
+import { AuthHook } from '../../../hooks/use-auth'
+import { Login } from '../login'
+import { Router, Route } from 'react-router-dom'
 
-import Routes from '../../constants/routes'
-import { Button, Level } from '@reapit/elements'
+jest.mock('../../../core/connect-session', () => ({
+  ReapitConnectBrowserSessionInstance: {
+    instance: {
+      connectLoginRedirect: jest.fn(),
+    },
+  },
+}))
 
-import * as loginStyles from './__styles__/styles'
+describe('Login', () => {
+  it('should match a snapshot', () => {
+    const history = createBrowserHistory()
+    const wrapper = render(
+      <AuthContext.Provider value={mockContext}>
+        <Router history={history}>
+          <Route>
+            <Login />
+          </Route>
+        </Router>
+      </AuthContext.Provider>,
+    )
+    expect(wrapper).toMatchSnapshot()
+  })
 
-import logoImage from '../../assets/images/reapit-graphic.jpg'
-import connectImage from '../../assets/images/reapit-connect.png'
-import { AuthContext } from '../../context'
-import { ReapitConnectBrowserSessionInstance } from '../../core/connect-session'
+  it('should match a snapshot', () => {
+    const wrapper = render(
+      <AuthContext.Provider value={{} as AuthHook}>
+        <Login />
+      </AuthContext.Provider>,
+    )
+    expect(wrapper).toMatchSnapshot()
+  })
+})
 
-export const Login: React.FunctionComponent = () => {
-  const { wrapper, container, image } = loginStyles
-  const { loginSession } = React.useContext(AuthContext)
+describe('loginHandler', () => {
+  it('should correctly call redirect on click', () => {
+    const wrapper = shallow(<Login />)
 
-  if (loginSession) {
-    return <Redirect to={Routes.HOME} />
-  }
+    wrapper
+      .find(Button)
+      .first()
+      .simulate('click')
 
-  return (
-    <div className={container}>
-      <div className={wrapper}>
-        <Level>
-          <img src={connectImage} alt="Reapit Connect Graphic" />
-        </Level>
-        <Level>
-          <Button
-            fullWidth
-            type="submit"
-            variant="primary"
-            onClick={ReapitConnectBrowserSessionInstance.instance.connectLoginRedirect}
-          >
-            Login
-          </Button>
-        </Level>
-      </div>
-
-      <div className={image}>
-        <img src={logoImage} alt="Reapit Graphic" />
-      </div>
-    </div>
-  )
-}
-
-export default Login
+    // FIXME: add check if it call with location.origin
+    expect(ReapitConnectBrowserSessionInstance.instance.connectLoginRedirect).toHaveBeenCalledTimes(1)
+  })
+})
